+++ packages/genre-tree-view/src/index.ts
@@ -1,7 +1,9 @@
 export { GenreTree } from "./GenreTree";
 export { GenreTreeWheel } from "./GenreTreeWheel";
+export { GenreTreeWheelRight } from "./GenreTreeWheelRight";
 export { getGenreTreeColor, WHEEL_DEFAULT_FRAME_HEIGHT } from "./constants";
 export { groupNodesByRoot } from "./root-grouping";
 export type { GenreTreeNode, GenreTreeProps, GenreTreePlayState, TreeOrientation } from "./types";
 export type { GenreTreeWheelProps } from "./GenreTreeWheel";
+export type { GenreTreeWheelRightProps } from "./GenreTreeWheelRight";
 export type { GenreTreeRootGroup } from "./root-grouping";